requirements.txt
streamlit>=1.28.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlite3-utils>=3.34
passlib>=1.7.4
bcrypt>=4.0.1